feat: AlertDelegate 프로토콜과 AlertManager 구현
